staketia migration - upgrade handler function (#1216)
