+++ x/staketia/types/celestia.go
@@ -16,5 +16,9 @@ const (
 
 	SafeAddressOnStride            = "stride18p7xg4hj2u3zpk0v9gq68pjyuuua5wa387sjjc" // S3
 	OperatorAddressOnStride        = "stride1ghhu67ttgmxrsyxljfl2tysyayswklvxs7pepw" // OP-STRIDE
-	CelestiaUnbondingPeriodSeconds = uint64(21 * 24 * 60 * 60)                       // 21 days
+	CelestiaUnbondingPeriodDays    = 21
+	CelestiaUnbondingPeriodSeconds = uint64(CelestiaUnbondingPeriodDays * 24 * 60 * 60) // 21 days
+
+	CelestiaConnectionId = "connection-125"
+	CelestiaBechPrefix   = "celestia"
 )
